Moving screen classes into UI package
In order to follow a better app structure I'm separating classes into packages related with its topic
