+++ KEAndroidApp/app/src/main/java/dev/chuby/ke_android_app/ui/ActionsFragment.kt
@@ -1,4 +1,4 @@
-package dev.chuby.ke_android_app
+package dev.chuby.ke_android_app.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import android.widget.Button
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
+import dev.chuby.ke_android_app.R
 
 class ActionsFragment : Fragment() {
