+++ KEAndroidApp/app/src/main/java/dev/chuby/ke_android_app/ui/AttendeeFragment.kt
@@ -1,15 +1,16 @@
-package dev.chuby.ke_android_app
+package dev.chuby.ke_android_app.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import dev.chuby.ke_android_app.R
 
-class AddAttendeeFragment : Fragment() {
+class AttendeeFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_add_attendee, container, false)
+        return inflater.inflate(R.layout.fragment_attendee, container, false)
     }
 
 }
