+++ KEAndroidApp/app/src/main/java/dev/chuby/ke_android_app/ui/AttendeesFragment.kt
@@ -1,10 +1,11 @@
-package dev.chuby.ke_android_app
+package dev.chuby.ke_android_app.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import dev.chuby.ke_android_app.R
 
 class AttendeesFragment : Fragment() {
